refactor: change types directories
types 폴더 디렉토리 구조 변경 및 api.ts 파일 분리

(cherry picked from commit 47332e722917801ca91cf4461c68857aaf75ff3b)
