+++ frontend/types/domains/applicationForms.ts
@@ -1,0 +1,15 @@
+export type Answer = {
+  contents: string;
+  recruitmentItemId: number;
+};
+
+export type ApplicationForm = {
+  id: number;
+  recruitmentId: number;
+  referenceUrl: string;
+  submitted: boolean;
+  answers: Answer[];
+  createdDateTime: string;
+  modifiedDateTime: string;
+  submittedDateTime: string | null;
+};
