+++ frontend/types/domains/user.ts
@@ -1,0 +1,9 @@
+export type User = {
+  id: number;
+  name: string;
+  email: string;
+  phoneNumber: string;
+  gender: "MALE" | "FEMALE";
+  birthday: Date;
+  password: string;
+};
