--- frontend/types/utility.ts
@@ -1,6 +1,0 @@
-export type RequestWithToken<T = {}> = { token: string } & T;
-
-export type ResponseDataWithMessage<T> = {
-  message: string;
-  body: T;
-};
